Break up evm result file
