+++ packages/arb-evm/evm/result.go
@@ -20,7 +20,6 @@ import (
 	"fmt"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/common/math"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/offchainlabs/arbitrum/packages/arb-evm/message"
 	"github.com/offchainlabs/arbitrum/packages/arb-util/common"
@@ -28,7 +27,6 @@ import (
 	"github.com/offchainlabs/arbitrum/packages/arb-util/value"
 	"github.com/pkg/errors"
 	"math/big"
-	"math/rand"
 )
 
 type ResultType int
@@ -47,173 +45,6 @@ const (
 
 type Result interface {
 	AsValue() value.Value
-}
-
-type Provenance struct {
-	L1SeqNum        *big.Int
-	ParentRequestId common.Hash
-	IndexInParent   *big.Int
-}
-
-func NewProvenanceFromValue(val value.Value) (Provenance, error) {
-	failRet := Provenance{}
-	tup, ok := val.(*value.TupleValue)
-	if !ok {
-		return failRet, errors.New("val must be a tuple")
-	}
-	if tup.Len() != 3 {
-		return failRet, errors.Errorf("expected tuple of length 3, but recieved tuple of length %v", tup.Len())
-	}
-
-	// Tuple size already verified above, so error can be ignored
-	l1SeqNumVal, _ := tup.GetByInt64(0)
-	parentRequestIdVal, _ := tup.GetByInt64(1)
-	indexInParentVal, _ := tup.GetByInt64(2)
-
-	l1SeqNumInt, ok := l1SeqNumVal.(value.IntValue)
-	if !ok {
-		return failRet, errors.New("provenance l1SeqNum must be an int")
-	}
-
-	parentRequestIdInt, ok := parentRequestIdVal.(value.IntValue)
-	if !ok {
-		return failRet, errors.New("provenance parentRequestId must be an int")
-	}
-
-	indexInParentInt, ok := indexInParentVal.(value.IntValue)
-	if !ok {
-		return failRet, errors.New("provenance indexInParent must be an int")
-	}
-
-	indexInParent := indexInParentInt.BigInt()
-	if indexInParent.Cmp(math.MaxBig256) == 0 {
-		indexInParent = nil
-	}
-
-	var parentRequestId common.Hash
-	copy(parentRequestId[:], math.U256Bytes(parentRequestIdInt.BigInt()))
-
-	return Provenance{
-		L1SeqNum:        l1SeqNumInt.BigInt(),
-		ParentRequestId: parentRequestId,
-		IndexInParent:   indexInParent,
-	}, nil
-}
-
-func (p Provenance) AsValue() value.Value {
-	parent := p.IndexInParent
-	if parent == nil {
-		parent = math.MaxBig256
-	}
-	// Static slice correct size, so error can be ignored
-	tup, _ := value.NewTupleFromSlice([]value.Value{
-		value.NewIntValue(p.L1SeqNum),
-		value.NewIntValue(new(big.Int).SetBytes(p.ParentRequestId[:])),
-		value.NewIntValue(parent),
-	})
-	return tup
-}
-
-type IncomingRequest struct {
-	Kind       inbox.Type
-	Sender     common.Address
-	MessageID  common.Hash
-	Data       []byte
-	ChainTime  inbox.ChainTime
-	Provenance Provenance
-}
-
-func NewIncomingRequestFromValue(val value.Value) (IncomingRequest, error) {
-	failRet := IncomingRequest{}
-	tup, ok := val.(*value.TupleValue)
-	if !ok {
-		return failRet, errors.New("val must be a tuple")
-	}
-	if tup.Len() != 7 {
-		return failRet, errors.Errorf("expected tuple of length 7, but recieved tuple of length %v", tup.Len())
-	}
-
-	// Tuple size already verified above, so error can be ignored
-	kind, _ := tup.GetByInt64(0)
-	blockNumber, _ := tup.GetByInt64(1)
-	timestamp, _ := tup.GetByInt64(2)
-	sender, _ := tup.GetByInt64(3)
-	inboxSeqNum, _ := tup.GetByInt64(4)
-	messageData, _ := tup.GetByInt64(5)
-	provenanceVal, _ := tup.GetByInt64(6)
-
-	kindInt, ok := kind.(value.IntValue)
-	if !ok {
-		return failRet, errors.New("inbox message kind must be an int")
-	}
-
-	blockNumberInt, ok := blockNumber.(value.IntValue)
-	if !ok {
-		return failRet, errors.New("blockNumber must be an int")
-	}
-
-	timestampInt, ok := timestamp.(value.IntValue)
-	if !ok {
-		return failRet, errors.New("timestamp must be an int")
-	}
-
-	senderInt, ok := sender.(value.IntValue)
-	if !ok {
-		return failRet, errors.New("sender must be an int")
-	}
-
-	messageIDInt, ok := inboxSeqNum.(value.IntValue)
-	if !ok {
-		return failRet, errors.New("inboxSeqNum must be an int")
-	}
-	var messageID common.Hash
-	copy(messageID[:], math.U256Bytes(messageIDInt.BigInt()))
-
-	data, err := inbox.ByteStackToHex(messageData)
-	if err != nil {
-		return failRet, errors.Wrap(err, "unmarshalling input data")
-	}
-
-	provenance, err := NewProvenanceFromValue(provenanceVal)
-	if err != nil {
-		return failRet, err
-	}
-
-	return IncomingRequest{
-		Kind:      inbox.Type(kindInt.BigInt().Uint64()),
-		Sender:    inbox.NewAddressFromInt(senderInt),
-		MessageID: messageID,
-		Data:      data,
-		ChainTime: inbox.ChainTime{
-			BlockNum:  common.NewTimeBlocks(blockNumberInt.BigInt()),
-			Timestamp: timestampInt.BigInt(),
-		},
-		Provenance: provenance,
-	}, nil
-}
-
-func NewRandomIncomingRequest() IncomingRequest {
-	return IncomingRequest{
-		Kind:      inbox.Type(rand.Uint32()),
-		Sender:    common.RandAddress(),
-		MessageID: common.RandHash(),
-		Data:      common.RandBytes(200),
-		ChainTime: inbox.NewRandomChainTime(),
-	}
-}
-
-func (ir IncomingRequest) AsValue() value.Value {
-	// Static slice correct size, so error can be ignored
-	tup, _ := value.NewTupleFromSlice([]value.Value{
-		value.NewInt64Value(int64(ir.Kind)),
-		value.NewIntValue(ir.ChainTime.BlockNum.AsInt()),
-		value.NewIntValue(ir.ChainTime.Timestamp),
-		inbox.NewIntFromAddress(ir.Sender),
-		value.NewIntValue(new(big.Int).SetBytes(ir.MessageID[:])),
-		inbox.BytesToByteStack(ir.Data),
-		ir.Provenance.AsValue(),
-	})
-	return tup
 }
 
 type TxResult struct {
@@ -404,137 +235,6 @@ func parseTxResult(l1MsgVal value.Value, resultInfo value.Value, gasInfo value.V
 	}, nil
 }
 
-type OutputStatistics struct {
-	GasUsed      *big.Int
-	TxCount      *big.Int
-	EVMLogCount  *big.Int
-	AVMLogCount  *big.Int
-	AVMSendCount *big.Int
-}
-
-func (os *OutputStatistics) AsValue() value.Value {
-	// Static slice correct size, so error can be ignored
-	tup, _ := value.NewTupleFromSlice([]value.Value{
-		value.NewIntValue(os.GasUsed),
-		value.NewIntValue(os.TxCount),
-		value.NewIntValue(os.EVMLogCount),
-		value.NewIntValue(os.AVMLogCount),
-		value.NewIntValue(os.AVMSendCount),
-	})
-	return tup
-}
-
-type BlockInfo struct {
-	BlockNum   *big.Int
-	Timestamp  *big.Int
-	GasLimit   *big.Int
-	BlockStats *OutputStatistics
-	ChainStats *OutputStatistics
-}
-
-func (b *BlockInfo) LastAVMLog() *big.Int {
-	return new(big.Int).Sub(b.ChainStats.AVMLogCount, big.NewInt(1))
-}
-
-func (b *BlockInfo) FirstAVMLog() *big.Int {
-	return new(big.Int).Sub(b.LastAVMLog(), b.BlockStats.AVMLogCount)
-}
-
-func (b *BlockInfo) LastAVMSend() *big.Int {
-	return new(big.Int).Sub(b.ChainStats.AVMSendCount, big.NewInt(1))
-}
-
-func (b *BlockInfo) FirstAVMSend() *big.Int {
-	return new(big.Int).Sub(b.LastAVMSend(), b.BlockStats.AVMSendCount)
-}
-
-func (b *BlockInfo) AsValue() value.Value {
-	// Static slice correct size, so error can be ignored
-	tup, _ := value.NewTupleFromSlice([]value.Value{
-		value.NewInt64Value(1),
-		value.NewIntValue(b.BlockNum),
-		value.NewIntValue(b.Timestamp),
-		value.NewIntValue(b.GasLimit),
-		b.BlockStats.AsValue(),
-		b.ChainStats.AsValue(),
-	})
-	return tup
-}
-
-func parseBlockResult(blockNum value.Value, timestamp value.Value, gasLimit value.Value, blockStatsRaw value.Value, chainStatsRaw value.Value) (*BlockInfo, error) {
-	blockNumInt, ok := blockNum.(value.IntValue)
-	if !ok {
-		return nil, errors.New("blockNum must be an int")
-	}
-	timestampInt, ok := timestamp.(value.IntValue)
-	if !ok {
-		return nil, errors.New("timestamp must be an int")
-	}
-	gasLimitInt, ok := gasLimit.(value.IntValue)
-	if !ok {
-		return nil, errors.New("gasLimit must be an int")
-	}
-	blockStats, err := parseOutputStatistics(blockStatsRaw)
-	if err != nil {
-		return nil, err
-	}
-
-	chainStats, err := parseOutputStatistics(chainStatsRaw)
-	if err != nil {
-		return nil, err
-	}
-
-	return &BlockInfo{
-		BlockNum:   blockNumInt.BigInt(),
-		Timestamp:  timestampInt.BigInt(),
-		GasLimit:   gasLimitInt.BigInt(),
-		BlockStats: blockStats,
-		ChainStats: chainStats,
-	}, nil
-}
-
-func parseOutputStatistics(val value.Value) (*OutputStatistics, error) {
-	tup, ok := val.(*value.TupleValue)
-	if !ok || tup.Len() != 5 {
-		return nil, errors.New("expected result to be nonempty tuple")
-	}
-
-	// Tuple size already verified above, so error can be ignored
-	gasUsed, _ := tup.GetByInt64(0)
-	txCount, _ := tup.GetByInt64(1)
-	evmLogCount, _ := tup.GetByInt64(2)
-	avmLogCount, _ := tup.GetByInt64(3)
-	avmSendCount, _ := tup.GetByInt64(4)
-
-	gasUsedInt, ok := gasUsed.(value.IntValue)
-	if !ok {
-		return nil, errors.New("gasUsed must be an int")
-	}
-	txCountInt, ok := txCount.(value.IntValue)
-	if !ok {
-		return nil, errors.New("txCount must be an int")
-	}
-	evmLogCountInt, ok := evmLogCount.(value.IntValue)
-	if !ok {
-		return nil, errors.New("evmLogCount must be an int")
-	}
-	avmLogCountInt, ok := avmLogCount.(value.IntValue)
-	if !ok {
-		return nil, errors.New("avmLogCount must be an int")
-	}
-	avmSendCountInt, ok := avmSendCount.(value.IntValue)
-	if !ok {
-		return nil, errors.New("avmSendCount must be an int")
-	}
-	return &OutputStatistics{
-		GasUsed:      gasUsedInt.BigInt(),
-		TxCount:      txCountInt.BigInt(),
-		EVMLogCount:  evmLogCountInt.BigInt(),
-		AVMLogCount:  avmLogCountInt.BigInt(),
-		AVMSendCount: avmSendCountInt.BigInt(),
-	}, nil
-}
-
 func NewResultFromValue(val value.Value) (Result, error) {
 	tup, ok := val.(*value.TupleValue)
 	if !ok || tup.Len() == 0 {
@@ -585,18 +285,6 @@ func NewTxResultFromValue(val value.Value) (*TxResult, error) {
 	txRes, ok := res.(*TxResult)
 	if !ok {
 		return nil, errors.New("got block result but expected transaction")
-	}
-	return txRes, nil
-}
-
-func NewBlockResultFromValue(val value.Value) (*BlockInfo, error) {
-	res, err := NewResultFromValue(val)
-	if err != nil {
-		return nil, err
-	}
-	txRes, ok := res.(*BlockInfo)
-	if !ok {
-		return nil, errors.New("got transaction result but expected block")
 	}
 	return txRes, nil
 }
